+++ app.py
@@ -1,11 +1,11 @@
-from flask import Flask
-
-app = Flask(__name__)
-
-@app.route('/')
-def hello():
-    return 'Hello, Semicolon Devs'
-
-@app.route('/ping')
-def ping():
-    return 'pong!'
+# from flask import Flask
+#
+# app = Flask(__name__)
+#
+# @app.route('/')
+# def hello():
+#     return 'Hello, Semicolon Devs'
+#
+# @app.route('/ping')
+# def ping():
+#     return 'pong!'
