+++ flaskr/__init__.py
@@ -1,0 +1,28 @@
+import os
+from flask import Flask
+
+def create_app(test_config=None):
+    app = Flask(__name__, instance_relative_config=True)
+    app.config.from_mapping(
+        SECRET_KEY = 'dev',
+        DATABASE=os.path.join(app.instance_path, 'flaskr.sqlite'),
+    )
+    if test_config is None:
+        app.config.from_pyfile('config.py', silent=True)
+    else:
+        app.config.from_mapping(test_config)
+
+    try:
+        os.makedirs(app.instance_path)
+    except OSError:
+        pass
+
+    @app.route('/')
+    def hello():
+        return 'Hello, Semicolon Devs'
+
+    @app.route('/ping')
+    def ping():
+        return 'pong!'
+
+    return app
